scc/requirements.txt
Click==7.0
Flask==1.1.1
Flask-API==1.1
itsdangerous==1.1.0
Jinja2==2.10.3
MarkupSafe==1.1.1
Pillow==6.2.1
pkg-resources==0.0.0
Werkzeug==0.16.0
Twython
